requirements.txt
django
mysqlclient
pillow
celery
redis